refactoring qna + search logic to a single endpoint
